LensApp V0.05
Updated a lot of the error handling so that the program would compile and not return infinite loops.
Not restricted to specific iFile functions, so I decided to use iFile.peak as a basis of looking at the data so that if the document is blank, you can add additional data. I do want some error handling better that you do not get prompted like that, but for now, it works